Move NewConfigFromEnv to test/dbutils
