+++ db/config.go
@@ -1,7 +1,5 @@
 package db
 
-import "os"
-
 // Config provide fields to configure the pool
 type Config struct {
 	// Database name
@@ -18,24 +16,4 @@ type Config struct {
 
 	// Port Number
 	Port string
-}
-
-// NewConfigFromEnv creates config from standard postgres environment variables,
-// see https://www.postgresql.org/docs/11/libpq-envars.html for details
-func NewConfigFromEnv() Config {
-	return Config{
-		Database: getEnv("PGDATABASE", "polygon-hermez"),
-		User:     getEnv("PGUSER", "hermez"),
-		Password: getEnv("PGPASSWORD", "polygon"),
-		Host:     getEnv("PGHOST", "localhost"),
-		Port:     getEnv("PGPORT", "5432"),
-	}
-}
-
-func getEnv(key string, defaultValue string) string {
-	value, exists := os.LookupEnv(key)
-	if exists {
-		return value
-	}
-	return defaultValue
 }
