+++ test/dbutils/dbutils.go
@@ -2,6 +2,7 @@ package dbutils
 
 import (
 	"context"
+	"os"
 
 	"github.com/hermeznetwork/hermez-core/db"
 )
@@ -32,4 +33,24 @@ func InitOrReset(cfg db.Config) error {
 		return err
 	}
 	return nil
+}
+
+// NewConfigFromEnv creates config from standard postgres environment variables,
+// see https://www.postgresql.org/docs/11/libpq-envars.html for details
+func NewConfigFromEnv() db.Config {
+	return db.Config{
+		Database: getEnv("PGDATABASE", "polygon-hermez"),
+		User:     getEnv("PGUSER", "hermez"),
+		Password: getEnv("PGPASSWORD", "polygon"),
+		Host:     getEnv("PGHOST", "localhost"),
+		Port:     getEnv("PGPORT", "5432"),
+	}
+}
+
+func getEnv(key string, defaultValue string) string {
+	value, exists := os.LookupEnv(key)
+	if exists {
+		return value
+	}
+	return defaultValue
 }
